+++ coauthors_test.go
@@ -9,27 +9,27 @@ func TestStartDoneCoAuthors(t *testing.T) {
 
 	setWorkingDir(tempDir + "/alice")
 	start(configuration)
-	createFile(t, "file3.txt", "owqe")
+	createFile(t, "file3.txt", "contentIrrelevant")
 	next(configuration)
 
 	setWorkingDir(tempDir + "/local")
 	start(configuration)
-	createFile(t, "file1.txt", "asdf")
+	createFile(t, "file1.txt", "contentIrrelevant")
 	next(configuration)
 
 	setWorkingDir(tempDir + "/localother")
 	start(configuration)
-	createFile(t, "file2.txt", "asdf")
+	createFile(t, "file2.txt", "contentIrrelevant")
 	next(configuration)
 
 	setWorkingDir(tempDir + "/alice")
 	start(configuration)
-	createFile(t, "file4.txt", "zcvx")
+	createFile(t, "file4.txt", "contentIrrelevant")
 	next(configuration)
 
 	setWorkingDir(tempDir + "/bob")
 	start(configuration)
-	createFile(t, "file5.txt", "oiuo")
+	createFile(t, "file5.txt", "contentIrrelevant")
 	next(configuration)
 
 	setWorkingDir(tempDir + "/local")