+++ mob.go
@@ -27,29 +27,47 @@ var (
 	Debug      = false // override with --debug parameter
 )
 
+type DoneSquash string
+
+const (
+	Squash    DoneSquash = "squash"
+	NoSquash             = "no-squash"
+	SquashWip            = "squash-wip"
+)
+
+func doneSquash(value string) DoneSquash {
+	switch value {
+	case "false", NoSquash:
+		return NoSquash
+	case SquashWip:
+		return SquashWip
+	}
+	return Squash
+}
+
 type Configuration struct {
-	CliName                        string // override with MOB_CLI_NAME
-	RemoteName                     string // override with MOB_REMOTE_NAME
-	WipCommitMessage               string // override with MOB_WIP_COMMIT_MESSAGE
-	GitHooksEnabled                bool   // override with MOB_GIT_HOOKS_ENABLED
-	RequireCommitMessage           bool   // override with MOB_REQUIRE_COMMIT_MESSAGE
-	VoiceCommand                   string // override with MOB_VOICE_COMMAND
-	VoiceMessage                   string // override with MOB_VOICE_MESSAGE
-	NotifyCommand                  string // override with MOB_NOTIFY_COMMAND
-	NotifyMessage                  string // override with MOB_NOTIFY_MESSAGE
-	NextStay                       bool   // override with MOB_NEXT_STAY
-	StartIncludeUncommittedChanges bool   // override with MOB_START_INCLUDE_UNCOMMITTED_CHANGES variable
-	StashName                      string // override with MOB_STASH_NAME
-	WipBranchQualifier             string // override with MOB_WIP_BRANCH_QUALIFIER
-	WipBranchQualifierSeparator    string // override with MOB_WIP_BRANCH_QUALIFIER_SEPARATOR
-	WipBranchPrefix                string // override with MOB_WIP_BRANCH_PREFIX
-	DoneSquash                     bool   // override with MOB_DONE_SQUASH
-	Timer                          string // override with MOB_TIMER
-	TimerRoom                      string // override with MOB_TIMER_ROOM
-	TimerLocal                     bool   // override with MOB_TIMER_LOCAL
-	TimerRoomUseWipBranchQualifier bool   // override with MOB_TIMER_ROOM_USE_WIP_BRANCH_QUALIFIER
-	TimerUser                      string // override with MOB_TIMER_USER
-	TimerUrl                       string // override with MOB_TIMER_URL
+	CliName                        string     // override with MOB_CLI_NAME
+	RemoteName                     string     // override with MOB_REMOTE_NAME
+	WipCommitMessage               string     // override with MOB_WIP_COMMIT_MESSAGE
+	GitHooksEnabled                bool       // override with MOB_GIT_HOOKS_ENABLED
+	RequireCommitMessage           bool       // override with MOB_REQUIRE_COMMIT_MESSAGE
+	VoiceCommand                   string     // override with MOB_VOICE_COMMAND
+	VoiceMessage                   string     // override with MOB_VOICE_MESSAGE
+	NotifyCommand                  string     // override with MOB_NOTIFY_COMMAND
+	NotifyMessage                  string     // override with MOB_NOTIFY_MESSAGE
+	NextStay                       bool       // override with MOB_NEXT_STAY
+	StartIncludeUncommittedChanges bool       // override with MOB_START_INCLUDE_UNCOMMITTED_CHANGES variable
+	StashName                      string     // override with MOB_STASH_NAME
+	WipBranchQualifier             string     // override with MOB_WIP_BRANCH_QUALIFIER
+	WipBranchQualifierSeparator    string     // override with MOB_WIP_BRANCH_QUALIFIER_SEPARATOR
+	WipBranchPrefix                string     // override with MOB_WIP_BRANCH_PREFIX
+	DoneSquash                     DoneSquash // override with MOB_DONE_SQUASH
+	Timer                          string     // override with MOB_TIMER
+	TimerRoom                      string     // override with MOB_TIMER_ROOM
+	TimerLocal                     bool       // override with MOB_TIMER_LOCAL
+	TimerRoomUseWipBranchQualifier bool       // override with MOB_TIMER_ROOM_USE_WIP_BRANCH_QUALIFIER
+	TimerUser                      string     // override with MOB_TIMER_USER
+	TimerUrl                       string     // override with MOB_TIMER_URL
 }
 
 func (c Configuration) wipBranchQualifierSuffix() string {
@@ -270,7 +288,7 @@ func getDefaultConfiguration() Configuration {
 		StartIncludeUncommittedChanges: false,
 		WipBranchQualifier:             "",
 		WipBranchQualifierSeparator:    "-",
-		DoneSquash:                     true,
+		DoneSquash:                     Squash,
 		Timer:                          "",
 		TimerLocal:                     true,
 		TimerRoom:                      "",
@@ -346,7 +364,7 @@ func parseUserConfiguration(configuration Configuration, path string) Configurat
 		case "MOB_WIP_BRANCH_QUALIFIER_SEPARATOR":
 			setUnquotedString(&configuration.WipBranchQualifierSeparator, key, value)
 		case "MOB_DONE_SQUASH":
-			setBoolean(&configuration.DoneSquash, key, value)
+			setMobDoneSquash(&configuration, key, value)
 		case "MOB_TIMER":
 			setUnquotedString(&configuration.Timer, key, value)
 		case "MOB_TIMER_ROOM":
@@ -419,7 +437,7 @@ func parseProjectConfiguration(configuration Configuration, path string) Configu
 		case "MOB_WIP_BRANCH_QUALIFIER_SEPARATOR":
 			setUnquotedString(&configuration.WipBranchQualifierSeparator, key, value)
 		case "MOB_DONE_SQUASH":
-			setBoolean(&configuration.DoneSquash, key, value)
+			setMobDoneSquash(&configuration, key, value)
 		case "MOB_TIMER":
 			setUnquotedString(&configuration.Timer, key, value)
 		case "MOB_TIMER_ROOM":
@@ -467,6 +485,20 @@ func setBoolean(s *bool, key string, value string) {
 	debugInfo("Overwriting " + key + " =" + strconv.FormatBool(boolValue))
 }
 
+func setMobDoneSquash(configuration *Configuration, key string, value string) {
+	boolValue, err := strconv.ParseBool(value)
+	if err != nil {
+		sayWarning("Could not set key from configuration file because value is not parseable (" + key + "=" + value + ")")
+		return
+	}
+	if boolValue {
+		configuration.DoneSquash = Squash
+	} else {
+		configuration.DoneSquash = NoSquash
+	}
+	debugInfo("Overwriting " + key + " =" + strconv.FormatBool(boolValue))
+}
+
 func parseEnvironmentVariables(configuration Configuration) Configuration {
 	setStringFromEnvVariable(&configuration.CliName, "MOB_CLI_NAME")
 	if configuration.CliName != getDefaultConfiguration().CliName {
@@ -497,7 +529,7 @@ func parseEnvironmentVariables(configuration Configuration) Configuration {
 
 	setBoolFromEnvVariable(&configuration.StartIncludeUncommittedChanges, "MOB_START_INCLUDE_UNCOMMITTED_CHANGES")
 
-	setBoolFromEnvVariable(&configuration.DoneSquash, "MOB_DONE_SQUASH")
+	setDoneSquashFromEnvVariable(&configuration, "MOB_DONE_SQUASH")
 
 	setStringFromEnvVariable(&configuration.Timer, "MOB_TIMER")
 	setStringFromEnvVariable(&configuration.TimerRoom, "MOB_TIMER_ROOM")
@@ -545,6 +577,22 @@ func setBoolFromEnvVariable(s *bool, key string) {
 	}
 }
 
+func setDoneSquashFromEnvVariable(configuration *Configuration, key string) {
+	value, set := os.LookupEnv(key)
+	if !set {
+		return
+	}
+
+	configuration.DoneSquash = doneSquash(value)
+
+	if value == "" {
+		debugInfo("ignoring " + key + "=" + value + " (empty string)")
+		return
+	}
+
+	debugInfo("overriding " + key + "=" + string(configuration.DoneSquash))
+}
+
 func removed(key string, message string) {
 	if _, set := os.LookupEnv(key); set {
 		say("Configuration option '" + key + "' is no longer used.")
@@ -581,7 +629,7 @@ func config(c Configuration) {
 	say("MOB_WIP_BRANCH_QUALIFIER" + "=" + quote(c.WipBranchQualifier))
 	say("MOB_WIP_BRANCH_QUALIFIER_SEPARATOR" + "=" + quote(c.WipBranchQualifierSeparator))
 	say("MOB_WIP_BRANCH_PREFIX" + "=" + quote(c.WipBranchPrefix))
-	say("MOB_DONE_SQUASH" + "=" + strconv.FormatBool(c.DoneSquash))
+	say("MOB_DONE_SQUASH" + "=" + string(c.DoneSquash))
 	say("MOB_TIMER" + "=" + quote(c.Timer))
 	say("MOB_TIMER_ROOM" + "=" + quote(c.TimerRoom))
 	say("MOB_TIMER_ROOM_USE_WIP_BRANCH_QUALIFIER" + "=" + strconv.FormatBool(c.TimerRoomUseWipBranchQualifier))
@@ -619,9 +667,11 @@ func parseArgs(args []string, configuration Configuration) (command string, para
 			}
 			i++ // skip consumed parameter
 		case "--squash":
-			newConfiguration.DoneSquash = true
+			newConfiguration.DoneSquash = Squash
 		case "--no-squash":
-			newConfiguration.DoneSquash = false
+			newConfiguration.DoneSquash = NoSquash
+		case "--squash-wip":
+			newConfiguration.DoneSquash = SquashWip
 		default:
 			if i == 1 {
 				command = arg
@@ -686,8 +736,6 @@ func execute(command string, parameter []string, configuration Configuration) {
 			squashWipGitEditor(parameter[1], configuration)
 		} else if len(parameter) > 1 && parameter[0] == "--git-sequence-editor" {
 			squashWipGitSequenceEditor(parameter[1], configuration)
-		} else {
-			squashWip(configuration)
 		}
 	case "version", "--version", "-v":
 		version()
@@ -819,8 +867,8 @@ func getMobTimerRoom(configuration Configuration) string {
 		currentBaseBranch, _ := determineBranches(currentBranch, gitBranches(), configuration)
 
 		if currentBranch.IsWipBranch(configuration) {
-			wipBranchWithouthWipPrefix := currentBranch.removeWipPrefix(configuration).Name
-			currentWipBranchQualifier = removePrefix(removePrefix(wipBranchWithouthWipPrefix, currentBaseBranch.Name), configuration.WipBranchQualifierSeparator)
+			wipBranchWithoutWipPrefix := currentBranch.removeWipPrefix(configuration).Name
+			currentWipBranchQualifier = removePrefix(removePrefix(wipBranchWithoutWipPrefix, currentBaseBranch.Name), configuration.WipBranchQualifierSeparator)
 		}
 	}
 
@@ -1176,6 +1224,10 @@ func done(configuration Configuration) {
 		return
 	}
 
+	if configuration.DoneSquash == SquashWip {
+		squashWip(configuration)
+	}
+
 	git("fetch", configuration.RemoteName, "--prune")
 
 	baseBranch, wipBranch := determineBranches(gitCurrentBranch(), gitBranches(), configuration)
@@ -1198,7 +1250,7 @@ func done(configuration Configuration) {
 
 		git("branch", "-D", wipBranch.Name)
 
-		if uncommittedChanges && !configuration.DoneSquash { // give the user the chance to name their final commit
+		if uncommittedChanges && configuration.DoneSquash != Squash { // give the user the chance to name their final commit
 			git("reset", "--soft", "HEAD^")
 		}
 
@@ -1216,7 +1268,7 @@ func done(configuration Configuration) {
 
 		if hasUncommittedChanges() {
 			sayTodo("To finish, use", "git commit")
-		} else if configuration.DoneSquash {
+		} else if configuration.DoneSquash == Squash {
 			sayInfo("nothing was done, so nothing to commit")
 		}
 
@@ -1236,7 +1288,7 @@ func gitRootDir() string {
 }
 
 func squashOrNoCommit(configuration Configuration) string {
-	if configuration.DoneSquash {
+	if configuration.DoneSquash == Squash {
 		return "--squash"
 	} else {
 		return "--no-commit"
@@ -1375,14 +1427,9 @@ Basic Commands(Options):
   done
     [--no-squash]                        Do not squash commits from wip branch
     [--squash]                           Squash commits from wip branch
+    [--squash-wip]                       Squash wip commits from wip branch, maintaining manual commits
   reset
     [--branch|-b <branch-postfix>]       Set wip branch to 'mob/<base-branch>/<branch-postfix>'
-
-Experimental Commands:
-  squash-wip                             Combines wip commits in wip branch with subsequent manual commits to leave only manual commits.
-                                         ! Works only if all wip commits have the same wip commit message !
-    [--git-editor]                       Not intended for manual use. Used as a non-interactive editor (GIT_EDITOR) for git.
-    [--git-sequence-editor]              Not intended for manual use. Used as a non-interactive sequence editor (GIT_SEQUENCE_EDITOR) for git.
 
 Timer Commands:
   timer <minutes>    start a <minutes> timer